Fix VM Abort when disassembling last rounds of those from recipe
Since you drop the last round you had for the selected recipe, it take it out of your inventory before it can find them on the ground, and if you no longer have any recipes, it has none to iterate through to find the selected one.

This way the filtering only happens when it's drawing to the screen, which is the only time it's needed anyways.  This way whatever's selected is stored until it's picked up again.