Add safety checks in canAssemble() and canDisassemble()
Hopefully this fixes the null pointer issue, since it's technically possible the recipe being passed in is null, which should just return false if that happens.

Also added a safety check for the owner of the item being present for assembly mode, since it's doing that in disassembly and because the URL should really only assemble if someone dropped it to begin with.